refactor(finyk): migrate DOM-free domain/lib into @sergeant/finyk-domain
Винесено чисту бізнес-логіку Фініка в новий workspace-пакет
packages/finyk-domain, щоб її без змін могли споживати apps/web та
майбутній apps/mobile (Expo). Web-специфічні частини (localStorage-
обгортки, finykStorage/finykBackup, storageManager, lsStats, demoData)
залишаються в apps/web.

Що зробив:
- Створив packages/finyk-domain (tsconfig, vitest, exports для
  ./constants, ./utils, ./domain(/*), ./lib(/*)).
- Перенiс apps/web/src/modules/finyk/{constants.ts, domain/*,
  lib/{accounts,categories,debt,formatting,goals,transactions,spending,
  recurringDetect,forecastEngine}.ts} + їх *.test.ts у пакет (git mv).
- Інлайніл THEME_HEX у debtEngine.ts, щоб пакет не тягнув web-only
  залежність @shared/lib/themeHex.
- Залишив thin-shim файли на оригінальних шляхах, що реекспортять із
  @sergeant/finyk-domain/* — всі існуючі імпорти (../constants,
  ../domain/*, @finyk/utils, @finyk/*) працюють без змін.
- utils.ts у apps/web додатково реекспортить ./lib/lsStats (web-only).
- apps/web: @sergeant/finyk-domain як workspace-залежність + paths у
  tsconfig.json.

Перевірка:
- pnpm lint: OK (всі 6 пакетів + check-imports)
- pnpm --filter @sergeant/finyk-domain typecheck: OK
- pnpm --filter @sergeant/web typecheck: OK (включно з sw.ts)
- pnpm --filter @sergeant/finyk-domain test: 107 tests passed (9 files)
- pnpm --filter @sergeant/web test: 713 tests passed (76 files)
- Поведінка вебу не змінюється — суто переміщення файлів + shims.
