+++ apps/web/tsconfig.json
@@ -12,6 +12,8 @@
         "../../packages/api-client/src/react/index.ts"
       ],
       "@sergeant/api-client/*": ["../../packages/api-client/src/*"],
+      "@sergeant/finyk-domain": ["../../packages/finyk-domain/src/index.ts"],
+      "@sergeant/finyk-domain/*": ["../../packages/finyk-domain/src/*"],
       "@shared/*": ["./src/shared/*"],
       "@finyk/*": ["./src/modules/finyk/*"],
       "@fizruk/*": ["./src/modules/fizruk/*"],
