+++ packages/finyk-domain/src/domain/categories.test.ts
