+++ packages/finyk-domain/src/domain/debtEngine.test.ts
