+++ packages/finyk-domain/src/lib/calcMonthlyNeeded.test.ts
