+++ packages/finyk-domain/src/lib/index.ts
@@ -1,0 +1,11 @@
+// DOM-free helper-и Фініка. Web-only частини (`lsStats`, `finykStorage`,
+// `finykBackup`, `storageManager`, `demoData`) лишаються в `apps/web`.
+export * from "./accounts.js";
+export * from "./categories.js";
+export * from "./debt.js";
+export * from "./formatting.js";
+export * from "./goals.js";
+export * from "./transactions.js";
+export * from "./spending.js";
+export * from "./recurringDetect.js";
+export * from "./forecastEngine.js";
